Do not create a segment if there's no profile to flush

+++ pkg/phlaredb/profile_store.go
@@ -224,16 +224,22 @@ func (s *profileStore) prepareFile(path string) (f *os.File, err error) {
 // profiles, including ones added since the start of the call, but only those
 // that were added before certain point (this call). The same for s.slice.
 func (s *profileStore) cutRowGroup(count int) (err error) {
-	path := filepath.Join(
-		s.path,
-		fmt.Sprintf("%s.%d%s", s.persister.Name(), s.rowsFlushed, block.ParquetSuffix),
-	)
 	// if cutRowGroup fails record it as failed segment
 	defer func() {
 		if err != nil {
 			s.metrics.writtenProfileSegments.WithLabelValues("failed").Inc()
 		}
 	}()
+
+	size := s.loadProfilesToFlush(count)
+	if len(s.flushBuffer) == 0 {
+		return nil
+	}
+
+	path := filepath.Join(
+		s.path,
+		fmt.Sprintf("%s.%d%s", s.persister.Name(), s.rowsFlushed, block.ParquetSuffix),
+	)
 	// Removes the file if it exists. This can happen if the previous
 	// ingestion attempt failed.
 	if err := os.Remove(path); err == nil {
@@ -242,11 +248,6 @@ func (s *profileStore) cutRowGroup(count int) (err error) {
 	f, err := s.prepareFile(path)
 	if err != nil {
 		return err
-	}
-
-	size := s.loadProfilesToFlush(count)
-	if len(s.flushBuffer) == 0 {
-		return nil
 	}
 
 	n, err := parquet.CopyRows(s.writer, schemav1.NewInMemoryProfilesRowReader(s.flushBuffer))
